some code style changes

+++ vue/components/toolsoftool/snap.vue
@@ -40,9 +40,7 @@
       }
     },
     destroyed() {
-      if (snapInteraction) {
-        mapService.removeInteraction(snapInteraction);
-      }
+      snapInteraction && mapService.removeInteraction(snapInteraction);
     }
   }
 </script>
